Resolves #46
Basically the methodology behind enforcing the double parameter use in Math.methods was to redundantly cast an expression on the double variables. This way it cannot use it has an integer parameter. The IDE will bother you for having this redundant expression, but hopefully I won't waste three hours of my life again.